+++ src/Resources/app/administration/src/component/nosto-job-status-badge/index.js
@@ -2,36 +2,7 @@
  * @sw-package innovation
  */
 
-import template from './nosto-job-status-badge.html.twig';
-
 const { Component } = Shopware;
 
 /** @private */
-Component.register('nosto-job-status-badge', {
-    template,
-
-    props: {
-        status: {
-            type: String,
-            required: true,
-        },
-    },
-
-    computed: {
-        additionalClass() {
-            return this.status === 'running' ? '--pulse' : '';
-        },
-
-        variant() {
-            switch (this.status) {
-                case 'error':
-                    return 'error';
-                case 'succeed':
-                case 'running':
-                    return 'success';
-                default:
-                    return '';
-            }
-        },
-    },
-});
+Component.register('nosto-job-status-badge', () => import('./component'));
